Cheat tweaks and rephrasings.
Changed the phrasing and functionality of the Score Cap cheat to be much more straightforward and clear: It now increases the score cap when enabled, rather than having to be disabled to increase it.

"More Buses" was renamed to "Add extra buses" for similar reasons but the functionality is the same.